+++ src/main/java/com/xiliulou/electricity/service/impl/FranchiseeUserInfoServiceImpl.java
@@ -79,7 +79,7 @@ public class FranchiseeUserInfoServiceImpl implements FranchiseeUserInfoService 
     }
 
     @Override
-    public Integer queryCountByFranchiseeId(Integer id) {
+    public Integer queryCountByFranchiseeId(Long id) {
         return franchiseeUserInfoMapper.selectCount(new LambdaQueryWrapper<FranchiseeUserInfo>().eq(FranchiseeUserInfo::getFranchiseeId,id).last("limit 0,1"));
     }
